+++ server/db/models/Game.js
@@ -1,0 +1,12 @@
+const db = require('../db')
+
+const Game = db.define('game', {
+  roundNumber: {
+    type: db.Sequelize.INTEGER
+  },
+  players: {
+    type: db.Sequelize.ARRAY(db.Sequelize.STRING)
+  },
+})
+
+module.exports = Game
