+++ server/db/models/Submission.js
@@ -1,0 +1,15 @@
+const db = require('../db')
+
+const Submission = db.define('submission', {
+  type: {
+    type: db.Sequelize.ENUM('drawing', 'guess')
+  },
+  guess: {
+    type: db.Sequelize.STRING
+  },
+  drawingUrl: {
+    type: db.Sequelize.STRING
+  },
+})
+
+module.exports = Submission
